requirements.txt
elasticsearch==2.2.0
pyzmq==15.2.0
scapy==2.3.2
urllib3==1.14
wheel==0.29.0
